list_open doesn't write if contents are unchanged
The additional space isn't bad for small files, and it's not important
for large files compared to the time saved when you're just reading the
file and not writing.

It's only bad when you open a large file and modify it, because we store
an extra copy.

This might be avoidable with a good hashing scheme (hash string before
splitting, hash joined output to see if we should write back)